refactor: drop always-running policy — one-shot post-install launch only
Employees keep tray Quit; nothing auto-respawns the app anymore.
- Windows: repeating watchdog task replaced by a trigger-less task the MSI
  registers and starts once, so a silent push still kills the old app
  (pre-kill CA) and brings the new one up after msiexec exits.
- mac: revert postinstall to KeepAlive SuccessfulExit=false and delete the
  launchd-managed marker plumbing.
